*) Bugfix for last commit
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@1506 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/plasma/dbImport/dbImportManager.java
@@ -46,7 +46,7 @@ public class dbImportManager {
         
         for (int i=0; i < activeCount; i++) {
             dbImporter currThread = (dbImporter) importThreads[i];
-            if (currThread.getJobID() == Integer.valueOf(jobID).intValue()) {
+            if (currThread.getJobID() == jobID) {
                 return currThread;
             }                    
         }
